code/requirements.txt
scikit-learn==0.24.2
numpy==1.21.0
opencv-python==4.5.3.56
torch==1.9.0
torchvision==0.10.0
matplotlib==3.4.2
openslide-python==1.1.2
scipy==1.7.0
xmltodict==0.12.0
pandas==1.3.0
tensorboardX==2.4